+++ src/main/java/kr/ai/nemo/auth/controller/OauthController.java
@@ -1,18 +1,18 @@
 package kr.ai.nemo.auth.controller;
 
 import jakarta.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import kr.ai.nemo.auth.dto.TokenRefreshResponse;
-import kr.ai.nemo.auth.exception.OAuthErrorCode;
-import kr.ai.nemo.auth.exception.OAuthException;
 import kr.ai.nemo.auth.service.OauthService;
+import kr.ai.nemo.common.UriGenerator;
 import kr.ai.nemo.common.exception.ApiResponse;
-import kr.ai.nemo.user.dto.UserLoginResponse;
+import kr.ai.nemo.common.exception.CustomException;
+import kr.ai.nemo.common.exception.ResponseCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -23,21 +23,16 @@ public class OauthController {
   private final OauthService oauthService;
 
   @GetMapping("/auth/kakao/callback")
-  public ResponseEntity<ApiResponse<UserLoginResponse>> kakaoLogin(
+  public void kakaoLogin(
       @RequestParam(value = "code", required = false) String code,
       @RequestParam(value = "error", required = false) String error,
-      @RequestParam(value = "error_description", required = false) String errorDescription,
       HttpServletResponse response) {
-
-    if (error != null) {
-      throw new OAuthException(OAuthErrorCode.KAKAO_AUTH_ERROR);
+    oauthService.handleKakaoCallback(code, error, response);
+    try {
+      response.sendRedirect(UriGenerator.login().toString());
+    } catch (IOException e) {
+      throw new CustomException(ResponseCode.REDIRECT_FAIL);
     }
-
-    if (code == null || code.isEmpty()) {
-      throw new OAuthException(OAuthErrorCode.CODE_MISSING);
-    }
-
-    return ResponseEntity.ok(ApiResponse.success(oauthService.loginWithKakao(code, response)));
   }
 
   @PostMapping("/api/v1/auth/token/refresh")
